Rename insights.ts to insight-classes.ts
